W2S no longer requires UCanvas
i was trying to think of a way that i could use imgui to draw instead of canvas but the old function needed canvas to work so itd have to hook the game render thread. looked through code and noticed that canvas is only used to get the display size. since, i've changed the code to use imgui to get the size.